+++ editorextensions/codebeamer-custom-ui/modal/src/pages/import/components/importer/Importer.tsx
@@ -10,6 +10,7 @@ import {
 } from '../../../../constants/cb-import-defaults';
 import { CodeBeamerItem } from '../../../../models/codebeamer-item.if';
 import { RootState } from '../../../../store/store';
+import { useImportedItems } from '../../../../hooks/useImportedItems';
 
 import './importer.css';
 
@@ -25,36 +26,31 @@ export default function Importer(props: {
 
 	const [loaded, setLoaded] = useState(0);
 
-	// const importedItems = useImportedItems();
+	const importedItems = useImportedItems();
 
 	/**
 	 * Produces the "main query string", which defines what should be imported.
-	 * Can and should be extended by what should NOT be imported
 	 */
 	const getMainQueryString = () => {
-		if (props.queryString) return props.queryString;
-		else
-			return `${cbqlString}${
-				props.items.length
-					? ' AND item.id IN (' + props.items.join(',') + ')'
-					: ''
-			}`;
+		const mainQuery = cbqlString;
+		const selectedItemsFilter = props.items.length
+			? ` AND item.id IN (${props.items.join(',')})`
+			: '';
+		const importedItemsFilter = importedItems.length
+			? ` AND item.id NOT IN (${importedItems
+					.map((i) => i.itemId)
+					.join(',')})`
+			: '';
+
+		if (props.queryString) {
+			return `${props.queryString}${importedItemsFilter}`;
+		} else {
+			return `${mainQuery}${selectedItemsFilter}${importedItemsFilter}`;
+		}
 	};
 
 	//* applies all currently active filters by using the stored cbqlString,
 	//* then further filters out only the selected items (or takes all of 'em)
-
-	// const { data, error, isLoading } = useGetItemsQuery({
-	// 	page: DEFAULT_RESULT_PAGE,
-	// 	pageSize: MAX_ITEMS_PER_IMPORT,
-	// 	queryString: `${getMainQueryString()}${
-	// 		importedItems.length
-	// 			? ' AND item.id NOT IN (' +
-	// 			  importedItems.map((i) => i.itemId) +
-	// 			  ')'
-	// 			: ''
-	// 	}`,
-	// });
 
 	const { data, error, isLoading } = useGetItemsQuery({
 		page: DEFAULT_RESULT_PAGE,